GPU support for applying functions that return multiplle types.
We need special support to perform a CUDAVectorApply(function, vector)
operation when the function may return different types (e.g.,
return 1.0*x if x<5 else x). We pass to the CUDA execution one output
vector for each possible return type (std::vector<byte*> output_vectors)
and an index vector initialized to 0xFF which is updated depending on
the output type. We write index[x] = k if f(vec[x]) returns the (k+1)-th
possible type. Upon return, we use index and output_vectors to merge the
outputs into a single FORA array with different types.

We also change how we wrap the NativeCFG of the function to be applied
to accomodate for multiple return types by replacing the 'let' statement
with a continuation statement for each return type that performs the
store operation to the appropriate output_vector and updates the index
vector.